+++ CHANGELOG.md
@@ -29,6 +29,7 @@ For a full diff see [`7afa59c...main`][7afa59c...main].
 * Used `TimeKeeper` instead of `SlowTestCollector` in `Subscriber\TestPreparedSubscriber` ([#25]), by [@localheinz]
 * Used `TimeKeeper` and `Collector\Collector` instead of `SlowTestCollector` in `Subscriber\TestPassedSubscriber` ([#26]), by [@localheinz]
 * Composed maximum duration into `SlowTest` ([#37]), by [@localheinz]
+* Rendered maximum duration in report created by `DefaultReporter` ([#38]), by [@localheinz]
 
 ### Removed
 
@@ -53,5 +54,6 @@ For a full diff see [`7afa59c...main`][7afa59c...main].
 [#34]: https://github.com/ergebnis/phpunit-slow-test-detector/pull/34
 [#36]: https://github.com/ergebnis/phpunit-slow-test-detector/pull/36
 [#37]: https://github.com/ergebnis/phpunit-slow-test-detector/pull/37
+[#38]: https://github.com/ergebnis/phpunit-slow-test-detector/pull/38
 
 [@localheinz]: https://github.com/localheinz